Initial check in - sketching out the functionality
Sketching out the functionality of the tester. For simplicity,
currentyl implemented in Clojure (because I think in parentheses) but
the "real" project will likely be implemented in a different language.

This checkin concentrates on the main functionality and tries to
explore a good API metaphor to build the tester before things get
serious.